[kt-*-only] Mute MavenMigrateTest
This test is broken for quite a while, it was fixed in master
by Dmitry Gridin in da28b680c98dd5a534f4709310c279b8d6d70557.
Discussed with Dmitry Gridin and agreed that it is difficult to
cherry-pick fix to the previous platforms, so we decided to just
mute this test i.e. fix the current behaviour

GitOrigin-RevId: 506db7a54f4311988dcc4b2a959c5300b777ac61