[kt-*-only] Mute MavenMigrateTest
This test is broken for quite a while, it was fixed in master
by Dmitry Gridin in da28b680c98dd5a534f4709310c279b8d6d70557.
Discussed with Dmitry Gridin and agreed that it is difficult to
cherry-pick fix to the previous platforms, so we decided to just
mute this test i.e. fix the current behaviour

GitOrigin-RevId: c43dabbfd439ef708574b4775bd8aae6a9304868