Restore binary compatibility with kotlin IDE plugin
Currently, for 221 platform, we have 2 versions of the Kotlin IDE plugin:
- bundled
- published to the plugin portal with the new Kotlin front-end version

In the e951c78dab79d584 method LessonContext.prepareSample was changed
(added default parameter) and this change is source-compatible, but
unfortunately it isn't a binary compatible change.
And we already have plugin `1.7.0-Beta` for 221 IDEA published to the plugin
portal, and it uses `prepareSample(String):Unit` method.

So, to restore binary compatibility @JvmOverloads was added.
This change isn't necessary for 222 platform, because for 222 right now we have
only bundled kotlin plugin

GitOrigin-RevId: 3ecd24709245ad72873380607f19b64443c4d0fa