[gradle] store idea module output separately IDEA-235250
Gradle allows to configure IDEA module output path. This path was stored twice in different parts of `ExternalSourceDirectorySet` entity, causing confusion.

This path should only be set to a module in case of non-delegated build, as some IDEA features perform checks of compilation output and in case of delegated build it may be plain wrong

GitOrigin-RevId: 2b0fd18683c3145abfbef16deb05b769a831c3b4