[workspace model] Fix replace by source ordering at restore references in matching IDs
Due to the changed order of tests execution VFU can have different hasCodes depending on the
time of creation and same for the FileInDirectory#fileNameId. The hashCode affects order of
replaceWithMatchedEntities map. Thus, it can be replaced with list because features of hash map
not used. With this change the order at step 2 `Traverse entities of replaceWith store` will be
 the same at step 5 `Restore references in matching ids`
 It fixes `FoldersImportingTest.testAddingExistingGeneratedSourcesWithCustomTargetDir` with
 options `-Dmaven.import.to.workspace.model=true -Dmaven.import.use.tree.import=true` if
 we will execute all tests from this class at once

GitOrigin-RevId: 53513fb547511bd63c172d1777d5e81d495e03b7