RUBY-24684 RUBY-24322 extend the `rsync` copying on WSL to allow to copy individual files also, not only directories.
Previously, we didn't properly cache the gem binaries like `/usr/bin/bundler` or ``/usr/bin/racc`, because in `WSLDistribution#copyFromWsl` it was assumed that we would only copy between directories. That's why the resulting stubs contained only empty directories.

Fixed the same way it was fixed for SSH in `RSyncUtil#download`. This is not ideal, since we essentially duplicate the logic. But this should be dealt with in RUBY-29814.

GitOrigin-RevId: 37a38b3e73e3805a7ec591c35859ae020da507f4