ML in SE: Compute ML features even if we're not sorting by ML (IDEA-294763)
If we compute the features regardless of ordering method (ML or not), we solve the bug because when the state will be reported, it will already contain  calculated features.

Otherwise (before this commit), final feature calculation takes place only after an element gets selected. At that point, it is too late, because features will not be calculated properly (e.g. file will already be opened, so isOpen will always be true).

GitOrigin-RevId: 37fa90fcb1cb50388a8b633b51badc1c3c0ae3a4