PY-40581: WSL: Do not copy editable package sources to ``remote_sources``.
All python packages from any remote SDK (WSL included) must be copied to ``remote_sources`` since indexing WSL (as eny remote SDK) from Windows may be slow.

This should be done each time package is installed/upgraded.

But when you have editable package, there is no reason nor possibility to copy it on each change.

WSL provides access to its filesystem using VFS (see RemoteTargetEnvironmentWithLocalVfs) and we use it.

``remote_sync.py`` reports editable package as root. But we know that it resides in module content root, so we exclude it from copying, but add it as simple mapping and content root
instead.

With this change editable packages resolved not to ``remote_sources`` but to module content
root instead.

GitOrigin-RevId: 1557950f0b1ba588e5ef7e6a767c3d9c1d85ee28