[workspace model] run synchronization with project configuration files before background activities from plugins
Since DelayedProjectSynchronizer was converted to a background activity, its execution doesn't delay loading of the project so there is no need to run it as the last one. Moreover, as this activity is needed to react on changes made in project configuration files outside of IDE (e.g. if user had pulled new commits or switched to a different branch from a command line), it's better to run it sooner, and don't delay it until all background activities from all plugins are executed. This also may help to mitigate disruption caused by a problem in a plugin's activity (see IDEA-295627).

GitOrigin-RevId: a9ba3291ac3904aafdf660d08d9c5c9b7ede2295