[kotlin] disable the IR evaluator by default
Right now users face a lot of issues with the current implementation of the IR evaluator. It is better to disable it by default for now, and then enable it back when all issues will be solved.

(cherry picked from commit 7ba732c13a4decfdcae76ec74eacfd3175c0984e)

IJ-CR-25445

GitOrigin-RevId: d69976fa20f230e022d69142dc1b5818c71369df