[platform] restore special base interface for backgroundPostStartupActivity extensions (IJ-CR-25260)
Such activities may be executed while indexes aren't ready, so it's better to make this clear by using the base interface which extends DumbAware. Now the both declarations of the extension point specify the proper interface, so if an extension from some plugin violates the requirement, the platform will report an error from ExtensionPointImpl::checkExtensionType and ignore the problematic extension, so problems like IDEA-295627 won't happen.

GitOrigin-RevId: ab99d69b96876acca3887bbd6860ecabe60cdc3b