TestKotlinArtifacts: manually download necessary Kotlin artifacts instead of reusing .idea/libraries
This commit changes logic but doesn't change any semantic.

I switched to a different way because I'm going to add bunch of other
kotlinc artifacts which are not used as dependencies in intellij monorepo
(such as kotlin-stdlib-js), it doesn't make sense to add those libraries
as a dependencies to the intellij monorepo because it will increase
indexing time for nothing and those won't be "real" intellij
dependencies, so let's migrate to a new scheme in this commit

Also `kotlinStdlibCommonSources` is fixed now.

GitOrigin-RevId: d137e01c733b634f3a6c8c0616611acdd2d6574c