Add new Git update method: Fetch all
In some scenarios this may be more useful than the previously implemented
Fetch Default method. For example, when working with a fork on GitHub,
it is often useful to have two remotes: origin (your fork) and upstream (the
original repo) to periodically fetch from (e.g. to rebase a PR branch on the
updated upstream branch). In this case fetching from the default remote
(which is usually origin) does little good, and Fetch All becomes useful.