[rdct] set different config/system folders when the process starts (RDCT-6)
These changes come from amp/cwm-settings-stuff5 branch (IJ-CR-26507). They introduce the ability to customize config/system paths via PathCustomizer in the platform,
and provide the implementation for CWM which set different values for these paths for each process, allow running several processes from the same installation in parallel.
Also, it listens for changes in the original config directory and reloads corresponding components when files under that folder are modified,
and updates files in the original config directory when settings are changed.

GitOrigin-RevId: 838a99d8789cfcf7b08c3823933806273b474aac