angular: fix unstable MoveTest
The problem was that we encoded reference for the current moving files, but didn't preserve AST collecting for it. So, if GC happens right after encoding references it breaks the refactorings.
The fix is to keep hard reference to the AST, to avoid GC-ing it.

(cherry picked from commit b73a475f1aab2e0ad5d4cc914a7c165f0f0ab701)

IJ-CR-25907

GitOrigin-RevId: 1ea658d754d70b48205f7198286ac822893185e1