jvmTarget 1.6 -> 1.8 fallback is needed not only 212-222 platforms
jvmTarget fallback is a hack because JPS wasn't unbundled & bundled JPS
is 1.7.0 compiler (jvmTarget 1.6 is dropped in Kotlin 1.7.0). We thought
that we wouldn't need this hack in 222 because we would unbundle Kotlin
JPS.

But unfortunatelly there are cases when we have to fallback to the
bundled JPS during Maven and Gradle import (because "Kotlin JPS unbundled
maven artifacts" were not published for milestone (beta, m1, m2, rc)
version of kotlin in [1.6.0, 1.7.0]).

That's why we have to bring back Alexander Udalov's jvmTarget fallback
fix into master.

GitOrigin-RevId: ce587368d297d5a6822baf192e3f6d36dc217430