avoid slow menu path calculations in action search
Problem: after several find usages action search becomes really slow due to Kotlin resolve in `ShowRecentFindUsagesGroup.getChildren`.

The original logic used `null` for `event` and the slow logic was avoided. This commit restores that for all actions registered in XML. `DefaultActionGroup.getChildren(null, manager)` is synchronized and final, so it's safe to call outside of RA and on any subclass.

GitOrigin-RevId: a263d5e30a93acfdd40682b44567b5daddfc435c