settings-sync: forcibly apply changes from cloud when initializing via "Take from Server": IDEA-292372
Before the fix, we at first logged current ide settings, and then applied changes from cloud as in the normal situation. This caused merging local and cloud settings, which was incorrect since they should be deliberately applied to the local state, overwriting existing settings.

To fix this, we handle the "Take from Server" situation differently from others:
1. log current settings and merge then to master.
2. apply server settings directly to master.
3. set 'ide' and 'cloud' pointers to the master position as well.

This commit also changes the way how initialization operations and executed: we call operations directly instead of creating fake events for them and then processing these events. That seems to clear the flow.

GitOrigin-RevId: 883028034fa6f3045a6c4b6bf4c4663baea6ecc8