Do not steal KEY_TYPED in PotemkinProgress.EventStealer
Stealing KEY_TYPED events causes intermittent typing failure for some characters,
for example, space. The cause is that it is bound to some actions (Expand/collapse,
Show quick preview), which causes the action processing mechanism to fire up,
which in turn starts stealing events with PotemkinProgress.EventStealer.
Then there's a race condition: if this action processing mechanism completes
before the KEY_TYPED event for the pressed space arrives, then space is typed
correctly. Otherwise, it is not typed, as the KEY_TYPED event is stolen.

This started to happen after 6bed082. Before that, it would only steal VK_ESCAPE.
As EventStealer apparently doesn't need KEY_TYPED events anyway, it seems that
the most robust solution to this issue is to simply never steal KEY_TYPED events
in the first place. This way KEY_PRESSED events will be stolen and processed,
but characters will still be typed as usual.