Kotlin auto-completion: don't show aliased Java types
Kotlin has many type aliases mirroring Java types, e.g. ArrayList,
IllegalArgumentException etc. When auto-completing, both the Kotlin
alias and the original Java type tend to pop up in random order,
causing unintended ugly imports of classes such as
java.lang.IllegalArgumentException and so on.

Deal with it in a way similar to how Java classes that should
not be used in Kotlin are handled: ignore them on the first auto-complete
attempt.

Not sure if fqNameUtil.kt is the right place to put the list of aliased classes,
as it would have been much better to simply ask the compiler, just
like it's done with the FqName.isJavaClassNotToBeUsedInKotlin() method,
but I haven't been able to find anything like that in the compiler API. I'm
not overly familiar with it, though, so I might've missed something.