RUBY-29673 add line comment whitespace to the prefix if we are starting a new comment line for the user to write on
If we are starting a new comment line, then the user probably wants the line comment to already have the whitespace so they may start writing immediately, rather than having to add the whitespace themselves. The original check for the newline is more intended to prevent trailing whitespace when block line commenting multiple lines.

GitOrigin-RevId: bd657cfae611cd5d34b503274ed09479ceb56e23