settings-sync: store the time when snapshot was created: IDEA-292399
To resolve conflicts properly, we need the date of creation of a settings snapshot: to identify which set settings was modified later and prefer it.

Thus, when the snapshot is collected to be sent to the server, we write the date to the snapshot. Then we serialize the date to the .metainfo/timestamp file, and send it to the server along with actual settings files. When the snapshot is received, the timestamp is deserialized from the file to the SettingsSnapshot object, and then it is written to Git.

The obvious way to store the timestamp in Git - is to use the author-date of a commit. But unfortunately JGit doesn't provide an API to modify the author-date. To overcome this we store the timestamp in a Git Note attached to the commit.

GitOrigin-RevId: 0430f47672fc16d4a9d8f7d59764349a4a98a923