Don't depend on delegated build enabled/disabled state during Kotlin JPS importing
IJ-CR-25955

It's not expected that users should re-import the project when they
enable "delegated build".

We depended on delegated build enabled/disabled state because:
- Don't show notifications about fallback to the bundled JPS
- Don't flush anything to kotlinc.xml

Now, we don't need to depend on delegated build enabled/disabled state
because we won't show notifications when
`shouldImportKotlinJpsPluginVersionFromExternalBuildSystem` is `false`.
Though we will show notifications for too old and too new Kotlins but I
think that it's fine.

And flushing anything to kotlinc.xml isn't harmful at all.

Related issue KTIJ-21852

GitOrigin-RevId: 37145fc79b66cb2bf45f44b4962f922d68fb0384