don't pass negative fraction to indicator
The indicator is initialized with `indeterminate=true` and `fraction=0.0`.
If `ProgressState#fraction` is less than `0.0`, then it means that the client never supplied a fraction update,
because the check inside `FlowProgressSink#fraction` guarantees that it's on;y possible to supply non-negative fraction.

Setting negative fraction to indicator generates a warning and sets indeterminate=false, which disables indeterminate
animation. It should've failed with an error, but it doesn't since the code is old.

To summarize: the bug is a typo, and after the change the code does what it was intended to do.

GitOrigin-RevId: 9e27520fc17524343486109fbe8ff72d2c1cdfde