settings-sync: check for updates before pushing: IDEA-296476
If there is a newer version of the settings snapshot on the server, we should update before pushing. To achieve this, we used to rely only on the server's rejection feature: together with the file, the client sends the version hash that the clients expects to be on the server; if the server has another version, then it means that the client had outdated information and needs to be updated; then such push gets rejected, and we update.

However, this seems to be not reliable enough. First of all, currently we don't persist the hash which means that it gets null after IDE restart. This can be fixed in various ways (persist the hash; check for updates if hash is unknown), but it seems that simply checking for updates every time before pushing is rather cheap and is quite easy to implement.

GitOrigin-RevId: 802646ba2a96644c728bfeca6cd9d04fabb51e92