[ConsoleView] Fix updating of filters in modal state
This fix is edited according to the problem described in IDEA-295794
Filters are updated before the view is attached to the hierarchy, so modality state for NonBlockingReadAction is computed incorrectly. Now it schedules update after component is attached to hierarchy.
This fix is required for `added-at` preview in UiInspector window.

(cherry picked from commit 7b59d109a049bd60f91ba8719b86e84790bee1d3)

IJ-CR-26222

GitOrigin-RevId: 921b02a78fa74db51fa18ec36b98edf9b99aae97