[ConsoleView] Fix updating of filters in modal state
This fix is edited according to the problem described in IDEA-295794
Filters are updated before the view is attached to the hierarchy, so modality state for NonBlockingReadAction is computed incorrectly. Now it schedules update after component is attached to hierarchy.
This fix is required for `added-at` preview in UiInspector window.

GitOrigin-RevId: 7b59d109a049bd60f91ba8719b86e84790bee1d3