Don't show "fallback to bundled JPS" notification when the build isn't delegated to JPS
IJ-CR-25955

This Notification is irrelevant for users who don't use JPS for the
compilation. Idealy, we should also show the notification once the user
switched to the "delegated to JPS" mode but I'm lazy to implement it
right now. I wish we used any reactive library in IDEA that would make
the implementation no-brainer.

This commit is partial revert of
37145fc79b66cb2bf45f44b4962f922d68fb0384. Partial because we still flush
the version into kotlinc.xml

GitOrigin-RevId: 4dc2f8deb3da1785321281caba6b4b71f6b8bc5b