settings-sync: remove the fast-forward merge optimization
It is not properly detected (master can be at the same position as ide, but cloud can still branch off an earlier commit, so a proper merge will be needed), and it complicates the logic. On the other hand, fast-forward situations are already identified and optimized inside JGit's merge method.

GitOrigin-RevId: 9c69fa25ac934bf7bbb34a813c9d9507d6661e5b