feature-suggester: disable sharing of its settings: IDEA-296458
The Feature Suggester stores the timestamps when the IDE was started, to show suggestions for an action only if the action wasn't used during the last 20 days. To identify the time when an action was used, the plugin looks into ActionsLocalSummary. Since the latter is not shared, there is no sense to share FeatureSuggesterSettings either.

OTOH sharing the FeatureSuggesterSettings leads to the situation when the settings get modified too frequently thus cluttering the settings sync history, and also may lead to merge conflicts as in IDEA-296458.

Since sharing is useless now, and since the Feature Suggester is not enabled in the release, we disable the roaming. Later we can return to this issue and decide how to make the necessary information shareable so that it is both useful and non-distractive.

(cherry picked from commit aaf282314b5d4c98a238929dd2e089141983142d)

IJ-CR-26343

GitOrigin-RevId: 77b736984c9baa8841793a5c08c86c2e4d3766e2