[kotlin FE10Binding] Fix stackoverflow in KtSymbolBasedDeclarationDescriptor#equals
Current implementation of equals|hashCode is not great.
Unfortunately we cannot make it better.

Method analyze() in old FE called at the beginning of the intention/inspection and after that PSI could be mutated,
and then some bindingContext.get() called. Maybe we should not support such intention/inspection in the Fe10Wrapper.

That means, that in Fe10Binding wrapper FIR session will be dropped in the middle of the work of the in inspection/intention.
It means, that KtSymbols also will re-created for the same entities, so we cannot really use the equals/hashCode on them.
Unfortunately, we don't have a stable signature or something like that for significant part of our KtSymbols.

The good news is that if there were no PSI modification, we can use KtSymbols, because corresponding cache will be
reset only on PSI modification.

GitOrigin-RevId: 1a99d9d04eb94f106a835156609d1e81d36c38db