[actions] disable `Refactor This` in popups (IDEA-296961)
it won't be shown if a shortcut would be assigned to some group and Refactor This would be added to the group as well; it would be present in custom quick lists though. Let's suppose that Refactor This is present in main menu for learning purposes only.

(cherry picked from commit 865d89671b3e38fdbe9f4bdda30a29898da08713)

IJ-CR-26625

GitOrigin-RevId: 4002bba041c2a963c1c9199d90cb2fd3f26a31ae