IDEA-285336/IDEA-276966 don't require uninitialized context in tasks passed into `ExecutorService#submit` and such
These tasks are usually wrapped into a `FutureTask`, which implements `Runnable`, and then they are wrapped the second
time when this `Runnable` gets passed to `Executor#execute`. Before the change we had to track tasks passed into
`Executor#execute` to avoid double wrapping, which was enforced with an assertion, the check was done by instanceof-ing
`FutureTask`, and it proved to be unreliable, because there could be other `FutureTask`/`RunnableFuture`
implementations. After the change:
- all `Callable`-s passed into `ExecutorService#sumbit` are wrapped, but the assertion is turned off inside it;
- the wrapped callable is then wrapped into `FutureTask` as per `AbstractExecutorService#newTaskFor`;
- all `Runnable`-s passed into `Executor#execute`/`Application#invokeLater` are wrapped with the assertion turned on,
this also includes `FutureTask`-s.

GitOrigin-RevId: 937bb260e60f756b0b9818d099aefc0f367c323c