[kotlin] Fix IdeReplExecutionTest
This commit is an almost revert of the 4ce6b0287b6bb400ce17ae429002e549715607a0 commit.
The only change here is that `else ->` branch is used to fix the compilation.

The commit was reverted, because in `kt-master` branch custom JPS version is used, and Relp is using
this version as well to execute commands, so the ReplOutputHandler should be changed once we change version of the bundled JPS

GitOrigin-RevId: 60c0de0860e15578c8f9352c53f6859ec689d737