[ExpUI] Register an iconMapper to override Python file type icon
Previously, the mapping file was put in a wrong place (not at the root of
a "resources" directory) and was not registered in any plugin descriptor.

However, under debug, the right icon was still displayed because pythonFile.svg
is also remapped in all-platform PlatformIconMappings.json. In actual installers
pythonFile.svg and PlatformIconMappings.json end up in different classloaders,
so this fallback doesn't work.

GitOrigin-RevId: ee361cd7d2d0158b0cc96bd652a5cb53bb7d88c5