PY-54712 Remove a redundant smart pointer from PyAddImportFix
It fixes the warning about a field blocking an intention preview, even though
the preview can't be shown for this quick fix in the first place, as it is
registered for the whole file.

However, the fix is generic enough so that it can be used in other places,
e.g. being registered for some references inside a file, where the preview
is offered.

GitOrigin-RevId: e46d5f90c8050858dfbd19835ac5fa4ba48d6fbf