drop fast-track for shortcuts
Originally, fast-track was an optimization for cases like arrow keys in kt files. Now as most actions in our codebase are not `ActionUpdateThread.OLD_EDT` and thus do not request slow data-context data on EDT we can finally drop it. And also resolve the fast-action-wins-over-slow problem that was introduced by it.

GitOrigin-RevId: 1169c1c7ba62bfe39f84d545c9846eff7f96e15b