IDEA-198442 Add an ability to load custom icons by URL
The problem is that 'iconCustomization' can store actionId or icon path as a reference for icon. This change adds an ability to store also the URL of icon. So icons from platform and plugins can be used as custom icons for actions. It is required because some icons is not used in the actions, so they can not be referenced in 'iconCustomization' without this change.

GitOrigin-RevId: eba55f9b5832d7efbf6c8cd3d9e55f97451c45b7