[kotlin] IdeKotlinVersion: use build number in the artifactVersion for preliminary artifacts
Previously version as 1.7.20-Beta-333 had artifactVersion `1.7.20-Beta`.
And some test are failing, because maven artifacts for such version has
version `1.7.20-Beta-333`, because it is preliminary artifacts.

For the release artifacts like `1.7.0-Beta-release-336` we currently having:
  - maven artifacts with the name `1.7.0-Beta` published to Maven central
  - maven artifacts with the name `1.7.0-Beta-release-336` published via nightly
     published to the space/kotlin-dev
  - space/kotlin-ide-plugin-dependencies contains `1.7.0-Beta-release-336` as well

P.s. after this commit when we switch state in Kotlin release configuration to RELEASE
(i.e. build number will be like `1.7.20-Beta-release-336`) some IDE tests will fail, because
IDE will use the `1.7.20-Beta` version fot artifacts and artifacts is not yet published
to the maven central. And that is expected.

P.p.s. Our IDE tests on buildserver are using the artifacts from
the `Compiler Dist and Maven Artifacts for IDE` configuration,
and this configuration contains build number no matter what, because this configuration
contains artifacts for IDE i.e. we have the same logic as with the
space/kotlin-ide-plugin-dependencies repo.

GitOrigin-RevId: 934769971b334485c476541714ac1659f3ebb1fe