explicitly pack `intellij.platform.testFramework` into `testFramework.jar`
Before this change `intellij.platform.testFramework` was handled by `layout.withModule(name)`
which infers the JAR name from the module name, and it was not obvious that the JAR name is indeed `testFramework.jar`,
because old module name was used, and `intellij.platform.testFramework` has `testFramework` as old name.

See `ModuleRenamingHistoryState` and `.idea/modules.xml`.

GitOrigin-RevId: ca1ac3bea86353d6d8fcdb739c5bf4b331adde28