Make PyBoundFunction not unstub a wrapped PyFunction upon construction
It was caused by function.getNode() in its constructor. The right way to make
it work with PSI stubs is to delegate completely to the wrapped PyFunction.
Just accepting a PyFunctionStub in its constructor is not enough, as a reference
to the original stub owner is saved inside PSI stub-managing logic, hence it's
restored on a different (original) element.

This synthetic element is not used often in our codebase, but this change will
prevent unstubbing problems if additional usages appear eventually.

I stumbled across it while experimenting with returning a PyBoundFunction from
PyResolveUtil.resolveQualifiedNameInScope for PY-53671.

GitOrigin-RevId: 6b205a7eb5dbe764d1b135ddda975574785b25e2