[workspace model codegen tests] ensure that EntityCodeGenTest doesn't change shared state
Project model for light project must not be changed in the test code because it's shared between tests, a new LightProjectDescriptor should be configured instead. Changes in global settings must be restored when test finishes to avoid problems in other tests. In order to execute test on TeamCity, its module must be added to dependencies of one of *.main modules.

GitOrigin-RevId: c9d5f03d2a26402359785f0d3059e9184753bf9c