[kotlin] move qualifier type check out of findExpression
Checking if expression qualifier is an object is the only place where smart analysis is used in findElementUtils, which prevents it from being moved to kotlin.base module. This commit moves the qualifier type check to findElement in elementSelectionUtils, since it is the only place where such check is required.

GitOrigin-RevId: 93918715ed91798c2a7f549bb5b7096f85826320