[kotlin] Make failure message more clear in tests (KTIJ-22015)
After moving code fragment compilation cache directly into the 'KotlinEvaluator' class, tests started to use the 'KotlinEvaluator' logger for evaluation diagnostics (see 'LogPropagator')'.

Coincidentally, the same logger was used in 'loadClassSafely()', so additional error messages appeared in the propagated messages.

It seems to be a good idea to know if the compiling evaluator failed in tests, so instead of hiding the new error message, I decided to provide a more understandable message.

GitOrigin-RevId: 1138c5efdf5a2fc05fb77e3c57843b0a062b4923