DS-3706 Fix freeze in NotebookCellLinesIntervalDataRule by making it a slow rule
Even though the rule is used in some UI things like cell collapsing, it is not easy to untie it from PSI due to different token providers for different PSI languages.

In case someone has to revert this commit, they'll have to find a way to juxtapose `NotebookCellLinesProvider` implementations for editors without `LanguageExtension`.

GitOrigin-RevId: 2e8133ee83cec46428f83577fe5a914585c58485