selection expand/contract performance
before this patch, when the user held down shift-right to expand the selection continuously, mySelectionMarker was created and added to the tree on each keypress, which eventually slowed down the typing as the number of range markers was significant enough for the insertion to be noticeable. after this patch: dispose old selectionMarker right before overwriting it with the new to amortize GC costs and avoid sudden freezes when GC had to remove many nodes from mySelectionMarkerTree, making it essentially one-node tree.

GitOrigin-RevId: 60a86d8ec4a97386acfd4656f598e2723104c0da