PY-55338 Fix "connection refused" error on connecting from debugger to IDE with Docker-based target interpreter on Linux
For Docker on Linux the server socket allocated on "localhost" cannot be reached from a Docker container. For the sever socket to be available it must be allocated on the broadcast address of the dedicated "docker0" network interface. This goal is achieved by resolving a free local port against the prepared target environment available before the Python debugger process creation.

GitOrigin-RevId: 143d329f9a65d0c2af112e7ecc1552089d718cb9