PY-55338 Fix "connection refused" error on connecting from console to IDE with Docker-based target interpreter on Linux
For Docker on Linux the server socket allocated on "localhost" cannot be reached from a Docker container. For the sever socket to be available it must be allocated on the broadcast address of the dedicated "docker0" network interface. This goal is achieved by resolving a free local port against the prepared target environment available before the Python debugger process creation.

GitOrigin-RevId: bea96fb8d6d23de66b5a68384fdd5769c0a42846