settings-sync: don't delete the file if it's roaming type is disabled: IDEA-298677
When the state of a PersistentStateComponent is being set up, StateStorageManagerImpl#createFileBasedStorage() deletes the storage file from StreamProviders (if there are any) if the roaming type of that component is DISABLED.

However, the SettingsSyncIdeMediator operates with files in the APP_CONFIG, so the file gets deleted not from some external repository specific to a settings sync, but from the APP_CONFIG as well.

To fix this, we don't delete a file if the roaming type of its component is DISABLED. We shouldn't operate this via the StreamProvider anyway.

GitOrigin-RevId: aba818454a476524134ef8a727e97a156c22dc2b