IDEA-299280 do not throw from `ProgressIndicatorUtils.checkCancelledEvenWithPCEDisabled` in non-cancellable section 2
In indicator mode, `NonCancelableIndicator` is installed as current indicator, and it does not throw from `checkCanceled`.
In Job mode, `Cancellation.computeInNonCancelableSection` does not replace the current indicator, and the exception was
caused because `checkCanceled` was actually called on original (not `NonCancelableIndicator`) cancelled indicator.

This change makes the behaviour consistent regardless of whatever is going on in the thread:
non-cancellable section must not throw PCE _at all_.
This change continues f225cd4bb25deb5fa03c37749c055158fb0dc817 and should be considered together with it.

GitOrigin-RevId: 9fe5e50683e73388476b715f74eeb18ffd5b8cae