Drop kotlin-reflect artifact from Kotlin plugin
Review: https://jetbrains.team/p/ij/reviews/94196

^KTIJ-22276 Fixed

In Kotlin compiler we are going to use kotlin-reflect which is the same
as in Kotlin plugin minimally supported IDEA (which is Android Studio).

We are going to migrate to "old reflect" in Kotlin in those commits:
https://jetbrains.team/p/kt/reviews/6753

Since the Kotlin compiler will depend on "old reflect" it's safe to use
IDEA's reflect in Kotlin plugin and compiler classpath.

This commit should fix FL-11828

GitOrigin-RevId: 4341c860073e7517d5e186f9ec18114f3f7987b2