Clean up and extend the Javadoc for Logger
The '<p/>' HTML tags were supposed to be closing '</p>' tags, which are optional in Javadoc.

In the paragraphs about the error methods, move the subject further to the front.

Document the different log formats used by 'debug', 'debugValues' and 'error'. Mention that this logging framework,
unlike many others, doesn't use placeholders in messages.

GitOrigin-RevId: 54da175391d0c0be205f7632a436f6eeff51e53a