cancel indicator `Job` with special `CancellationException`
Before the change a Job was failed with PCE. The job had no parent, so the failure didn't affect it.
This change makes the behaviour more in line with coroutine semantics (cancellation vs failure),
and it's now possible to use this function with an arbitrary Job.
After the change it's also possible to catch this particular exception instance to handle accordingly.

GitOrigin-RevId: a2a357e2bf0a27f0f8821516360b1aa0e51920c8