settings-sync: remember the server version id only after applying the server version: IDEA-300592
There was an important variable CloudConfigServerCommunicator#knownVersionOfSnapshotZip which was storing the id of the settings sync file stored on the server, which version was known to the IDE. That variable is used by the push() process: the version id is sent to the server together with the snapshot, and the server checks it: if this version is the same as on the server, that means that the IDE already knows about the actual settings file residing on the server, the IDE has already received and applied it, and now it is ready to push a new version. OTOH if the version is different, that means that the IDE doesn't know about the latest version of the file (e.g. it was pushed from another client), therefore such push gets rejected requiring the IDE to update and merge the latest file version first (and perform another push attempt - after merging and with an actual version id).

There is also SettingsSyncUpdateChecker#scheduleUpdateFromServer which receives the settings snapshot from the server. It can be called from a thread other than the main sync activity in the SettingsSyncBridge, so both processes can execute simultaneously. But scheduleUpdateFromServer not only received the snapshot from the server: during that phase it also updated the knownVersionOfSnapshotZip.

It led to a race: if the scheduleUpdateFromServer received a snapshot and - more importantly - updated knownVersionOfSnapshotZip during the sync process, then the push caused by that sync process was using this version id from the variable, effectively overwriting anything which is written on the server instead of merging it. Or - as in the problem described in the ticket in the case of initialization - vice versa, overwriting the local change by the update happening after the push: if the push would be rejected due to unknown version id, the IDE would take the server state, and merge it with the local instead of overwriting the latter.

There are several possible solutions to the problem. For instance, probably an update shouldn't happen in a separate thread from other sync activities - that is probably going to be fixed later. But the most important part is that we shall not update the knownVersionOfSnapshotZip just because we've received a snapshot from the server: it should be updated only after that snapshot has been applied, recorded to the SettingsLog and merged with the local changes. Therefore, this variable can be set only from the SettingsSyncBridge.

As a side effect of the fix, the known server id is now persisted, so no update on start of every IDE session is needed anymore.

GitOrigin-RevId: 9ae7481f8ab01884a5c9aa7401273ab23c70a019