settings-sync: fix the GitSettingsLogTest setup: copy existing settings and set all branches to the same position
Initially, copying existing settings was done as a part of GitSettingsLog.initialize(). It changed in d6ae15f5, but tests were not adjusted. They didn't fail but they were working on an incorrect initial data.

The 'cloud' branch position was never adjusted, but again, no test relied on that, so that incorrect behavior was in place.

This commit setups the log correctly by copying all existing settings and putting all 3 branches in sync.

GitOrigin-RevId: a02b74fa2bbb12df3ea48a7c90fdb5ae85729468