settings-sync: fix GitSettingsLog.setBranchPosition: it should change the working copy as well
GitSettingsLog.set[Ide|Cloud]Position is intended to move the branch pointer to the new position. Although, it is not expected to change the working tree anyhow, and although it is incorrect to expect that (and it will be fixed separately), it is still unexpected that the working tree gets modified when the pointer is moved.

 Previously, only the branch reference was updated. Moreover, even the updateRef command could fail without isForceUpdate, because it checks if the content has been merged (which is not true in most cases).

 This commit both fixes the call to the updateRef command, and also resets the working tree to the new state. It fixes IDEA-300758.

GitOrigin-RevId: 436b1c7901be9c351b0543bdd0e1c68d91bbcabb