settings-sync: make sure that SettingsLog.collectCurrentSnapshot gets content from the master branch
Previously, collectCurrentSnapshot used just to return the current state of files on disk. While we consider that the 'master' branch is the only source of truth, during the sync procedure it might happen that a different branch gets checked out (e.g. to applyCloudState). Although it is probably not a correct flow, and it should be fixed, we still don't want the state of other branches be sent as the current settings snapshot to the clients: it always should be from the master branch.

This fixes IDEA-300758.

GitOrigin-RevId: 3a7f7678469e52901c7e38cfba4b6ae281a2a0d8