RUBY-30206 do not use the `YAMLKeysIndex` to get the fixture record's labels.
The changes to `YAMLKeysIndex` that we made in order for it to start indexing our YAML fixture files, seem to be causing a lot of problems in other products. So are switching to collecting the record labels by walking the PSI tree.

This commit also reverts changes made in e10523136684445cd9539cedb892266b5d1acfc8 ("RUBY-4056 adjust `YAMLKeysIndex`, so that it would also index rails fixtures.").

GitOrigin-RevId: 929855f914ecfe0fe2920dc4bc6597c8313bc9d2