`PlatformTaskSupportKt#showModalIndicator`: properly handle focus in glass pane based modal progress
Before the change `launch` was scheduled with `Dispatchers.EDT`, which delegates to `LaterInvocator`.
`LaterInvocator` has its own internal queue, so bare `SwingUtilities#invokeLater` inside `IdeGlassPaneImp.addImpl`
was not able to finish before the focus was requested on `ui.cancelButton`.
Changing the dispatcher to `RawSwingDispatcher` ensures the proper order of events.

This change also moves the focus routine after `yield()`, because it's not needed in heavy dialogs,
as they automatically get the focus.

GitOrigin-RevId: 95b234128a1560e9b38f757e7695e73845182eb9