extract `com.intellij.openapi.progress.CoroutinesKt#runUnderIndicator` to allow arbitrary indicators
`com.intellij.openapi.wm.ex.ProgressIndicatorEx` and implementations are not available from `intellij.platform.core`,
but it's necessary to be able to run older code under `ProgressIndicatorEx` which expects it,
e.g. unconditionally casting to it, like `CodeSmellDetector#findCodeSmells`.
It's up to the indicator implementation to forward text/text2/fraction updates to the context `ProgressSink`.

GitOrigin-RevId: 913b9b774f3125a5eec0327b2e81d114670b76d3