Consider all compiler args when determining --preview status
IntelliJ discovers whether a Gradle project should have preview features
enabled by checking if --enable-preview is added to the Java compile
arguments.

In Gradle 4.5, a new mechanism of adding compiler arguments was added in
the CommandLineArgumentProvider interface. This also changes the actual
accessor to be 'allCompilerArgs' rather than 'compilerArgs'.

Right now, if a CommandLineArgumentProvider adds --enable-preview (this
is done [here](https://github.com/palantir/gradle-baseline) then
IntelliJ will not discover this, leading to preview features not being
enabled.

This PR fixes the problem in a way that doesn't affect IntelliJ's gradle
compatibility (by checking if the new approach is possible and if not,
using the old one) and modifies the tests.