`runBlockingModal`: cancel `processEventQueueJob` then post an empty runnable to unblock `getNextEvent`
`invokeOnCompletion` is used instead of `finally` block, because in order to execute `finally`,
`processEventQueueJob` must yield first, but it's blocked in `getNextEvent`.
`invokeOnCompletion` is executed by whatever thread completed the job, in this case a `Dispatchers.Default` thread.

This change also fixes `coroutineContext` usage, which made the task job a child of `runBlocking`.

GitOrigin-RevId: 8767dd670b61799ecc00e9f590fc9534e5d92308