[platform] Take ModificationTracker objects into account in ListStoredProperty not only BaseState
Since not all classes can extend BaseState (e.g. it is needed for ProjectGroup, but why should it extend BaseState), it is better to use interface that provides modification counter. Already existed ModificationTracker is the best candidate for that.
Also, previously ListStoredProperty code didn't work properly for case when list values were changed, but modificationCount wasn't updated, because it wasn't changed structurally
For sure, it is better not to modify objects inside collections, but sometimes this is the only way...

Related issue: IDEA-298050

GitOrigin-RevId: 7bbbc979129acd64add1846fb9891c68c902415b