settings-sync: fix enabling/disabling of several plugins at once: IDEA-300962
Previously, we were iterating over the state of plugins, and enabling or disabling them immediately. But disabling a plugin led to the disabledListener to execute. And that listener updates the component state according to the actual state of the IDE, effectively settings the state of second plugin (which should be deleted but was waiting for its turn) back to enabled (because it was still enabled in the IDE).

In this commit we collect data to enable/disable/install plugins first, and only then we process it. If the state gets changed by a side effect (e.g. from the disabledListener) it will be processed later in its turn.

GitOrigin-RevId: 42a7a91fa49e31f129f46263db4903615a4d718e