[KTIJ-22685] Script as project entity: visualize dependencies
This commit introduces new custom entity for the WorkspaceModel - KotlinScriptEntity.
The idea is to have an explicit (+visible) representation of a script and
its dependencies. As of now the new entity is just a view of the underlying data
structure - ScriptClassRootsCache, controlled by ScriptConfigurationManager.

From now on every .kts file dependencies are put under `Project Structure` =>
`Libraries` menu. There are two types of them: `Script: <path>` and
`Script: path: sdk<name>` (in case project sdk differs from the one that script
requires).

All changes in this and subsequent commits of the series are made under
registry key `kotlin.scripts.as.entities`.

GitOrigin-RevId: 6af672d9dd138a8a9bd2e16e11f6dea6b2002211