[platform] IDEA-300530 Fix Active Ticket Changelist TODO Hide button
The cause was a simple typo of the copy-paste kind: a wrong component
was set as the preferred focusable component, which later lead to a
failed isShowing() check, which prevented the Hide action from being
performed.

In the case of the New UI it led to an NPE instead, which I haven’t
investigated, but after this fix it no longer happens either, so it’s
safe to assume the cause was the same, just the bug manifested itself
differently for the old and New UIs.

GitOrigin-RevId: 9d4cf3c547e9310fae7e3acd69d8e13cd2f0b047