[newUI] IDEA-292613 Don't show tooltip with context menu
Perhaps not the most elegant way to fix it, but a robust one.
Disable tooltips by setting the new client property on the
target component whenever the context menu is displayed.

A client property is used because there is no good place to
keep this flag: we can't place it in the component's class
because the menu doesn't know the exact class of the target
component (as it should be); and we can't place it in the
tooltip itself because the tooltip in theory could change
while the context menu is displayed.

One thing to note is that tool tips for other tool window
buttons are still shown. This may or may not be a good thing.

GitOrigin-RevId: 8753ed8c987d819dbb49ea01a19012c87314c628