settings-sync: don't load SettingsSyncPluginManager state from disk
We don't need to load the state from disk: on startup we update the component state by the state of the IDE (which plugins are enabled/disabled/installed). After that we listen to plugin changes, and update the state accordingly. If plugins change on another client, we can apply this new state right away, instead of writing it to disk and then loading it back from disk.

On the other hand, loading the state from disk led to issues like IDEA-300764, and requires synchronization, especially since loadState can be called in any moment.

GitOrigin-RevId: ffb220f0defe5a6c79e0d034c985654058ef3c38