[Workspace Model] Initialize bridges in separate handlers
At the moment, bridges are initialized in the workspace listener. This is a problem because if some other client wants to access the bridge of let's say new created module, they can't do it because the bridge will be available in the "next" version of the storage.

Since the bridges are important part of the IntelliJ, the solution is to move bridge initialization into special handlers that are called right before workspace model update and the builder is still available.

The problem with that approach is in the `subscribeAfterModuleLoading` listener.
If someone subscribes to this listener, listener collects all the events that were happen before modules loading, and executes them at once after the loading happens.
The described approach won't work with this type of listener.

After some analyzing it turned out that there is a single case when we have some events before modules are loaded - loading without cache. So, we've made a decision to move `loadModules` code a few lines up, so it's called in the moment of the project update, while builder is available.
After this change, the hacky `subscribeAfterModuleLoading` listener won't be needed.

At the moment, I've introduced a bit hairy code with a `loadModules` function in lambda, however it looks like it would be possible to get rid of `loadModules` in case of loading without cache at all, since the activities in this method are almost the same, as for any new added `ModuleEntity`. So, instead of calling this method, we can just create `ModuleEntity`es and this code will be executed as for regular module entities.

GitOrigin-RevId: fd5d8535c0839d2da6e62344758bbb624b9a96d0